Cleaning up the folder
type: chore

The goal of this update is to surface older project for others
to explore and use as a foundation for their learning. Cleaning
out the old checkpoint folders from Jupyter and adding in a
dedicated data file makes it a little neater.

At the end of this update, I am hoping to better explain some
fundemental data exploration steps such as dataset dimensions,
visualizations, and simple descriptive statistics. Also this
is a great intro to importing one's own data and applying the
same sort of analysis.